+++ src/interpolation.rs
@@ -5,11 +5,12 @@
 /// Available kind of interpolations.
 ///
 /// Feel free to visit each variant for more documentation.
+#[cfg(feature = "bezier")]
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 #[cfg_attr(feature = "serialization", derive(Deserialize, Serialize))]
 #[cfg_attr(feature = "serialization", serde(rename_all = "snake_case"))]
 pub enum Interpolation<T, V> {
-  /// Hold a [`Key<T, _>`] until the sampling value passes the normalized step threshold, in which
+  /// Hold a [`Key`] until the sampling value passes the normalized step threshold, in which
   /// case the next key is used.
   ///
   /// > Note: if you set the threshold to `0.5`, the first key will be used until half the time
@@ -17,7 +18,7 @@ pub enum Interpolation<T, V> {
   /// > first key will be kept until the next key. Set it to `0.` and the first key will never be
   /// > used.
   ///
-  /// [`Key<T, _>`]: crate::key::Key
+  /// [`Key`]: crate::key::Key
   Step(T),
   /// Linear interpolation between a key and the next one.
   Linear,
@@ -41,11 +42,35 @@ pub enum Interpolation<T, V> {
   ///   interpolation_ and it kicks ass too, but a bit less than cubic.
   #[cfg(feature = "bezier")]
   Bezier(V),
-  #[cfg(not(any(feature = "bezier")))]
-  #[doc(hidden)]
-  _V(std::marker::PhantomData<V>),
 }
 
+/// Available kind of interpolations.
+///
+/// Feel free to visit each variant for more documentation.
+#[cfg(not(feature = "bezier"))]
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "serialization", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serialization", serde(rename_all = "snake_case"))]
+pub enum Interpolation<T> {
+  /// Hold a [`Key`] until the sampling value passes the normalized step threshold, in which
+  /// case the next key is used.
+  ///
+  /// > Note: if you set the threshold to `0.5`, the first key will be used until half the time
+  /// > between the two keys; the second key will be in used afterwards. If you set it to `1.0`, the
+  /// > first key will be kept until the next key. Set it to `0.` and the first key will never be
+  /// > used.
+  ///
+  /// [`Key`]: crate::key::Key
+  Step(T),
+  /// Linear interpolation between a key and the next one.
+  Linear,
+  /// Cosine interpolation between a key and the next one.
+  Cosine,
+  /// Catmull-Rom interpolation, performing a cubic Hermite interpolation using four keys.
+  CatmullRom,
+}
+
+#[cfg(feature = "bezier")]
 impl<T, V> Default for Interpolation<T, V> {
   /// [`Interpolation::Linear`] is the default.
   fn default() -> Self {
@@ -53,3 +78,10 @@ impl<T, V> Default for Interpolation<T, V> {
   }
 }
 
+#[cfg(not(feature = "bezier"))]
+impl<T> Default for Interpolation<T> {
+  /// [`Interpolation::Linear`] is the default.
+  fn default() -> Self {
+    Interpolation::Linear
+  }
+}