+++ CHANGELOG.md
@@ -1,8 +1,11 @@
-# 1.1
+# 1.1.1
 
 > Mon Sep 22rd 2019
 
-- Yanked.
+- Add support for [Bézier curves](https://en.wikipedia.org/wiki/B%C3%A9zier_curve). This is
+  normally a breaking change so it’s currently disabled by default and available via the
+  `"bezier"` feature-gate.
+- Add `Spline::get`, `Spline::get_mut` and `Spline::replace`.
 
 # 1.0